@@ -1,7 +1,8 @@
 library flutter_list_bloc;
 
-export 'src/continuous_listview_bloc_builder.dart';
+export 'src/continuous_scroll_list_cubit_provider.dart';
+export 'src/list_bloc_builder.dart';
 export 'src/listview_bloc_builder.dart';
-export 'src/listview_paginated_bloc_builder.dart';
-export 'src/pageview_bloc_builder.dart';
+export 'src/page_view_bloc_builder.dart';
 export 'src/sliver_list_bloc_builder.dart';
+export 'src/types.dart';
