@@ -1,72 +1,44 @@
-import 'package:flutter/material.dart' hide Page;
-import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:flutter/material.dart';
 import 'package:list_bloc/list_bloc.dart';
 
-typedef Widget StateBuilder<T, F>(
-  BuildContext context,
-  Data<List<T>, F> state,
-);
-
-typedef Widget StateItemBuilder<T, F>(
-  BuildContext context,
-  Data<List<T>, F> state,
-  int index,
-  T item,
-);
+import 'list_bloc_builder.dart';
+import 'types.dart';
 
 class SliverListBlocBuilder<T, F> extends StatelessWidget {
-  final ListCubit<T, F>? cubit;
-
-  final StateBuilder<T, F> loadingBuilder;
-  final StateItemBuilder<T, F> itemBuilder;
-  final StateBuilder<T, F> emptyBuilder;
-  final StateBuilder<T, F>? headerBuilder;
-
   final ListCubit<T, F> Function(BuildContext context)? create;
+  final ListCubit<T, F>? cubit;
+  final ListItemBuilder<T, F> itemBuilder;
+  final ListStateBuilder<T, F> emptyBuilder;
+  final ListStateBuilder<T, F> loadingBuilder;
+  final ListStateBuilder<T, F>? headerBuilder;
   final int loadingItemsCount;
 
   const SliverListBlocBuilder({
-    this.cubit,
+    required this.emptyBuilder,
     required this.itemBuilder,
     required this.loadingBuilder,
-    required this.emptyBuilder,
     this.create,
-    this.loadingItemsCount = 3,
+    this.cubit,
     this.headerBuilder,
+    this.loadingItemsCount = 3,
   }) : assert((cubit != null) != (create != null));
 
   @override
-  Widget build(BuildContext context) {
-    final child = BlocBuilder<ListCubit<T, F>, Data<List<T>, F>>(
-      bloc: cubit,
-      builder: (context, state) {
-        final isEmpty = state is! Loading && (state.data?.length ?? 0) == 0;
-        final itemCount = (headerBuilder != null ? 1 : 0) +
-            (isEmpty
-                ? 1
-                : state is Loading
-                    ? (state.data?.length ?? 0) + loadingItemsCount
-                    : (state.data?.length ?? 0));
-
-        final sliver = SliverList(
+  Widget build(BuildContext context) => ListBlocBuilder(
+        emptyBuilder: emptyBuilder,
+        itemBuilder: itemBuilder,
+        loadingBuilder: loadingBuilder,
+        create: create,
+        cubit: cubit,
+        headerBuilder: headerBuilder,
+        loadingItemsCount: loadingItemsCount,
+        builder: (BuildContext context, Data<List<T>, F> state, int itemCount,
+                Widget Function(BuildContext, int) itemBuilder) =>
+            SliverList(
           delegate: SliverChildBuilderDelegate(
-            (context, index) {
-              if (headerBuilder != null && index == 0) {
-                return headerBuilder!(context, state);
-              }
-              final i = index - (headerBuilder != null ? 1 : 0);
-              if (isEmpty) return emptyBuilder(context, state);
-              if (state is Loading && i >= itemCount - loadingItemsCount - 1) {
-                return loadingBuilder(context, state);
-              }
-              return itemBuilder(context, state, i, state.data![i]!);
-            },
+            itemBuilder,
             childCount: itemCount,
           ),
-        );
-        return sliver;
-      },
-    );
-    return create != null ? BlocProvider(create: create!, child: child) : child;
-  }
+        ),
+      );
 }
